diploma_store/requirements.txt
Django==2.2
pytz==2019.1
sqlparse==0.3.0
django-ckeditor==5.7.1
